Enhancement: Rename Collector to Collector\DefaultCollector and extract Collector\Collector interface
